iterate on IntUiBridgeMenu
